# openapi.yaml
openapi: 3.1.0
...
info:
  title: LAGALia — Gestió de contractació pública
  version: 0.1.0
  description: |
    API de la plataforma de gestió de contractació pública de l'Ajuntament de Cunit.

    **API first**: tota funcionalitat del sistema és accessible per aquesta API.
    La interfície web n'és un client més, igual que els agents d'IA, n8n o
    qualsevol integració de tercers.

    Convencions generals a `docs/05-api.md`.
  contact:
    name: Ajuntament de Cunit
  license:
    name: EUPL-1.2
    identifier: EUPL-1.2
servers:
  - url: /api/v1
    description: Servidor actual
...
security:
  - bearerAuth: []
paths:
...
  /sync-runs:
    get:
      tags: [sync]
      operationId: listSyncRuns
      summary: Historial de sincronitzacions (keyset per id desc)
      x-required-scopes: [sync:read]
      parameters:
        - name: page[size]
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: page[cursor]
          in: query
          schema: { type: string }
        - name: filter[kind]
          in: query
          schema: { type: string, enum: [contracts, minor, cpv, extensions, enrichment, execution] }
        - name: filter[status]
          in: query
          schema: { type: string, enum: [running, success, failed, partial] }
      responses:
        "200":
          description: Runs
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/SyncRun" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /sync-runs/{id}/items:
    get:
      tags: [sync]
      operationId: listSyncRunItems
      summary: Detall per registre problemàtic d'un run
      x-required-scopes: [sync:read]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: page[size]
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: page[cursor]
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Item logs
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/SyncItem" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /sync-runs/actions/trigger:
    post:
      tags: [sync]
      operationId: triggerSync
      summary: Llança una sincronització (encua el job; el progrés es consulta a /jobs/{id})
      x-required-scopes: [sync:execute]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind]
              properties:
                kind:
                  type: string
                  enum: [contracts, minor, cpv, extensions, enrichment, execution]
                full:
                  type: boolean
                  default: false
                  description: Sync completa (ignora l'incremental); a enrichment, re-enriquir també els ja enriquits.
                limit:
                  type: [integer, "null"]
                  minimum: 1
                  maximum: 10000
                  description: Només enrichment; limita el nombre d'expedients.
      responses:
        "202":
          description: Job encuat
          content:
            application/json:
              schema:
                type: object
                required: [job_id, job_type]
                properties:
                  job_id: { type: string, format: uuid }
                  job_type: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }


  # ───────────────────────── public-registry ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token d'accés d'usuari (JWT curt) o clau de compte de servei (`sk_...`)
        per a agents d'IA i integracions. Sempre a la capçalera `Authorization`;
        mai per query string.
...
  responses:
...
    Unauthorized:
      description: Cal autenticar-se o la sessió ha caducat
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Autenticat però sense permís per a aquesta acció o recurs
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Recurs inexistent o fora del teu abast
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: L'acció entra en conflicte amb l'estat actual del recurs
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationError:
      description: Dades d'entrada no vàlides
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      description: Error segons RFC 9457. Mai inclou el cos de la petició original.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, examples: ["https://lagalia/errors/validation"] }
        title: { type: string, examples: ["Dades no vàlides"] }
        status: { type: integer, examples: [422] }
        detail: { type: string }
        instance: { type: string }
        trace_id:
          type: string
          description: Identificador de traça per a suport.
        errors:
          type: array
          description: Errors per camp, quan aplica.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    PageMeta:
      type: object
      required: [total]
      properties:
        total: { type: integer }
        next_cursor:
          type: [string, "null"]
...
    SyncRun:
      type: object
      required: [id, kind, trigger, status, new_count, updated_count, unchanged_count]
      properties:
        id: { type: integer, format: int64 }
        kind: { type: string, enum: [contracts, minor, cpv, extensions, enrichment, execution] }
        trigger: { type: string, enum: [manual, scheduled, api] }
        status: { type: string, enum: [running, success, failed, partial] }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        new_count: { type: integer }
        updated_count: { type: integer }
        unchanged_count: { type: integer }
        total_source: { type: [integer, "null"] }
        endpoint: { type: [string, "null"] }
        error_summary: { type: [object, "null"], additionalProperties: true }
    SyncItem:
      type: object
      required: [id, created_at]
      properties:
        id: { type: integer, format: int64 }
        file_code: { type: [string, "null"] }
        outcome: { type: [string, "null"] }
        message: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
...